[013] - Add ChartBar component and fixed wrong filenames of the last two components (Chart, ChartBar)
