+++ src/components/Chart/Chart.js
@@ -1,0 +1,20 @@
+import React from "react";
+import ChartBar from "./ChartBar";
+import "./Chart.css";
+
+const Chart = (props) => {
+  return (
+    <div className="chart">
+      {props.dataPoints.map((dataPoint) => (
+        <ChartBar
+          key={dataPoint.label}
+          value={dataPoint.value}
+          maxValue={null}
+          label={dataPoint.label}
+        />
+      ))}
+    </div>
+  );
+};
+
+export default Chart;
