+++ src/components/Chart/ChartBar.js
@@ -1,20 +1,21 @@
 import React from "react";
-import ChartBar from "./ChartBar";
-import "./Chart.css";
 
-const Chart = (props) => {
+import "./ChartBar.css";
+
+const ChartBar = (props) => {
+    let barFillHeight = '0%';
+
+    if (props.maxValue > 0) {
+        barFillHeight = Math.round((props.value / props.maxValue) * 100) + "%";
+    }
   return (
-    <div className="chart">
-      {props.dataPoints.map((dataPoint) => (
-        <ChartBar
-          key={dataPoint.label}
-          value={dataPoint.value}
-          maxValue={null}
-          label={dataPoint.label}
-        />
-      ))}
+    <div className="chart-bar">
+      <div className="chart-bar__inner">
+        <div className="chart-bar__fill" style={{height: barFillHeight}}></div>
+      </div>
+      <div className="chart-bar__label">{props.label}</div>
     </div>
   );
 };
 
-export default Chart;
+export default ChartBar;
